Chore: move types to types/snowpack
