+++ packages/snowpack/src/plugins/plugin-esbuild.ts
@@ -2,7 +2,7 @@ import {Service, startService} from 'esbuild';
 import * as colors from 'kleur/colors';
 import path from 'path';
 import {promises as fs} from 'fs';
-import {SnowpackPlugin, SnowpackConfig} from '../config';
+import {SnowpackPlugin, SnowpackConfig} from '../types/snowpack';
 
 let esbuildService: Service | null = null;
