+++ packages/snowpack/src/rollup-plugins/rollup-plugin-wrap-install-targets.ts
@@ -1,7 +1,7 @@
 import * as colors from 'kleur/colors';
 import path from 'path';
 import {Plugin} from 'rollup';
-import {InstallTarget} from '../scan-imports';
+import {InstallTarget} from '../types/snowpack';
 
 function autoDetectExports(fileLoc: string): string[] | undefined {
   try {
